@@ -8,14 +8,14 @@ use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
 
 class BaseTestCase extends WebTestCase
 {
-    static protected function createKernel(array $options = array())
+    protected static function createKernel(array $options = array())
     {
-        $config = isset($options['config']) ? $options['config'] : 'default.yml';
+        $config = $options['config'] ?? 'default.yml';
 
         return new AppKernel($config);
     }
 
-    protected final function importDatabaseSchema()
+    final protected function importDatabaseSchema()
     {
         foreach (self::$kernel->getContainer()->get('doctrine')->getManagers() as $em) {
             $this->importSchemaForEm($em);
