@@ -34,9 +34,9 @@ class Configuration implements ConfigurationInterface
      */
     public function getConfigTreeBuilder()
     {
-        $treeBuilder = new TreeBuilder();
-        $rootNode = $treeBuilder->root('jms_job_queue');
+        $treeBuilder = new TreeBuilder('jms_job_queue');
 
+        $rootNode = $treeBuilder->getRootNode();
         $rootNode
             ->children()
                 ->booleanNode('statistics')->defaultTrue()->end();
@@ -57,7 +57,7 @@ class Configuration implements ConfigurationInterface
         return $treeBuilder;
     }
 
-    private function addQueueOptions(ArrayNodeDefinition $def)
+    private function addQueueOptions(ArrayNodeDefinition $def): void
     {
         $def
             ->children()
