@@ -23,38 +23,39 @@ use PHPUnit\Framework\TestCase;
 
 class JobTest extends TestCase
 {
-    public function testConstruct()
+    public function testConstruct(): Job
     {
         $job = new Job('a:b', array('a', 'b', 'c'));
 
-        $this->assertEquals('a:b', $job->getCommand());
-        $this->assertEquals(array('a', 'b', 'c'), $job->getArgs());
-        $this->assertNotNull($job->getCreatedAt());
-        $this->assertEquals('pending', $job->getState());
-        $this->assertNull($job->getStartedAt());
+        self::assertEquals('a:b', $job->getCommand());
+        self::assertEquals(array('a', 'b', 'c'), $job->getArgs());
+        self::assertNotNull($job->getCreatedAt());
+        self::assertEquals('pending', $job->getState());
+        self::assertNull($job->getStartedAt());
 
         return $job;
     }
 
     /**
      * @depends testConstruct
-     * @expectedException JMS\JobQueueBundle\Exception\InvalidStateTransitionException
+     *
      */
-    public function testInvalidTransition(Job $job)
+    public function testInvalidTransition(Job $job): void
     {
+        $this->expectException(\JMS\JobQueueBundle\Exception\InvalidStateTransitionException::class);
         $job->setState('failed');
     }
 
     /**
      * @depends testConstruct
      */
-    public function testStateToRunning(Job $job)
+    public function testStateToRunning(Job $job): Job
     {
         $job->setState('running');
-        $this->assertEquals('running', $job->getState());
-        $this->assertNotNull($startedAt = $job->getStartedAt());
+        self::assertEquals('running', $job->getState());
+        self::assertNotNull($startedAt = $job->getStartedAt());
         $job->setState('running');
-        $this->assertSame($startedAt, $job->getStartedAt());
+        self::assertSame($startedAt, $job->getStartedAt());
 
         return $job;
     }
@@ -62,110 +63,108 @@ class JobTest extends TestCase
     /**
      * @depends testStateToRunning
      */
-    public function testStateToFailed(Job $job)
+    public function testStateToFailed(Job $job): void
     {
         $job = clone $job;
         $job->setState('running');
         $job->setState('failed');
-        $this->assertEquals('failed', $job->getState());
+        self::assertEquals('failed', $job->getState());
     }
 
     /**
      * @depends testStateToRunning
      */
-    public function testStateToTerminated(Job $job)
+    public function testStateToTerminated(Job $job): void
     {
         $job = clone $job;
         $job->setState('running');
         $job->setState('terminated');
-        $this->assertEquals('terminated', $job->getState());
+        self::assertEquals('terminated', $job->getState());
     }
 
     /**
      * @depends testStateToRunning
      */
-    public function testStateToFinished(Job $job)
+    public function testStateToFinished(Job $job): void
     {
         $job = clone $job;
         $job->setState('running');
         $job->setState('finished');
-        $this->assertEquals('finished', $job->getState());
+        self::assertEquals('finished', $job->getState());
     }
 
-    public function testAddOutput()
+    public function testAddOutput(): void
     {
         $job = new Job('foo');
-        $this->assertNull($job->getOutput());
+        self::assertNull($job->getOutput());
         $job->addOutput('foo');
-        $this->assertEquals('foo', $job->getOutput());
+        self::assertEquals('foo', $job->getOutput());
         $job->addOutput('bar');
-        $this->assertEquals('foobar', $job->getOutput());
+        self::assertEquals('foobar', $job->getOutput());
     }
 
-    public function testAddErrorOutput()
+    public function testAddErrorOutput(): void
     {
         $job = new Job('foo');
-        $this->assertNull($job->getErrorOutput());
+        self::assertNull($job->getErrorOutput());
         $job->addErrorOutput('foo');
-        $this->assertEquals('foo', $job->getErrorOutput());
+        self::assertEquals('foo', $job->getErrorOutput());
         $job->addErrorOutput('bar');
-        $this->assertEquals('foobar', $job->getErrorOutput());
+        self::assertEquals('foobar', $job->getErrorOutput());
     }
 
-    public function testSetOutput()
+    public function testSetOutput(): void
     {
         $job = new Job('foo');
-        $this->assertNull($job->getOutput());
+        self::assertNull($job->getOutput());
         $job->setOutput('foo');
-        $this->assertEquals('foo', $job->getOutput());
+        self::assertEquals('foo', $job->getOutput());
         $job->setOutput('bar');
-        $this->assertEquals('bar', $job->getOutput());
+        self::assertEquals('bar', $job->getOutput());
     }
 
-    public function testSetErrorOutput()
+    public function testSetErrorOutput(): void
     {
         $job = new Job('foo');
-        $this->assertNull($job->getErrorOutput());
+        self::assertNull($job->getErrorOutput());
         $job->setErrorOutput('foo');
-        $this->assertEquals('foo', $job->getErrorOutput());
+        self::assertEquals('foo', $job->getErrorOutput());
         $job->setErrorOutput('bar');
-        $this->assertEquals('bar', $job->getErrorOutput());
+        self::assertEquals('bar', $job->getErrorOutput());
     }
 
-    public function testAddDependency()
+    public function testAddDependency(): void
     {
         $a = new Job('a');
         $b = new Job('b');
-        $this->assertCount(0, $a->getDependencies());
-        $this->assertCount(0, $b->getDependencies());
+        self::assertCount(0, $a->getDependencies());
+        self::assertCount(0, $b->getDependencies());
 
         $a->addDependency($b);
-        $this->assertCount(1, $a->getDependencies());
-        $this->assertCount(0, $b->getDependencies());
-        $this->assertSame($b, $a->getDependencies()->first());
+        self::assertCount(1, $a->getDependencies());
+        self::assertCount(0, $b->getDependencies());
+        self::assertSame($b, $a->getDependencies()->first());
     }
 
-    /**
-     * @expectedException \LogicException
-     * @expectedExceptionMessage You cannot add dependencies to a job which might have been started already.
-     */
-    public function testAddDependencyToRunningJob()
+    public function testAddDependencyToRunningJob(): void
     {
+        $this->expectExceptionMessage("You cannot add dependencies to a job which might have been started already.");
+        $this->expectException(\LogicException::class);
         $job = new Job('a');
         $job->setState(Job::STATE_RUNNING);
         $this->setField($job, 'id', 1);
         $job->addDependency(new Job('b'));
     }
 
-    public function testAddRetryJob()
+    public function testAddRetryJob(): Job
     {
         $a = new Job('a');
         $a->setState(Job::STATE_RUNNING);
         $b = new Job('b');
         $a->addRetryJob($b);
 
-        $this->assertCount(1, $a->getRetryJobs());
-        $this->assertSame($b, $a->getRetryJobs()->get(0));
+        self::assertCount(1, $a->getRetryJobs());
+        self::assertSame($b, $a->getRetryJobs()->get(0));
 
         return $a;
     }
@@ -173,79 +172,79 @@ class JobTest extends TestCase
     /**
      * @depends testAddRetryJob
      */
-    public function testIsRetryJob(Job $a)
+    public function testIsRetryJob(Job $a): void
     {
-        $this->assertFalse($a->isRetryJob());
-        $this->assertTrue($a->getRetryJobs()->get(0)->isRetryJob());
+        self::assertFalse($a->isRetryJob());
+        self::assertTrue($a->getRetryJobs()->get(0)->isRetryJob());
     }
 
     /**
      * @depends testAddRetryJob
      */
-    public function testGetOriginalJob(Job $a)
+    public function testGetOriginalJob(Job $a): void
     {
-        $this->assertSame($a, $a->getOriginalJob());
-        $this->assertSame($a, $a->getRetryJobs()->get(0)->getOriginalJob());
+        self::assertSame($a, $a->getOriginalJob());
+        self::assertSame($a, $a->getRetryJobs()->get(0)->getOriginalJob());
     }
 
-    public function testCheckedAt()
+    public function testCheckedAt(): void
     {
         $job = new Job('a');
-        $this->assertNull($job->getCheckedAt());
+        self::assertNull($job->getCheckedAt());
 
         $job->checked();
-        $this->assertInstanceOf('DateTime', $checkedAtA = $job->getCheckedAt());
+        self::assertInstanceOf('DateTime', $checkedAtA = $job->getCheckedAt());
 
         $job->checked();
-        $this->assertInstanceOf('DateTime', $checkedAtB = $job->getCheckedAt());
-        $this->assertNotSame($checkedAtA, $checkedAtB);
+        self::assertInstanceOf('DateTime', $checkedAtB = $job->getCheckedAt());
+        self::assertNotSame($checkedAtA, $checkedAtB);
     }
 
-    public function testSameDependencyIsNotAddedTwice()
+    public function testSameDependencyIsNotAddedTwice(): void
     {
         $a = new Job('a');
         $b = new Job('b');
 
-        $this->assertCount(0, $a->getDependencies());
+        self::assertCount(0, $a->getDependencies());
         $a->addDependency($b);
-        $this->assertCount(1, $a->getDependencies());
+        self::assertCount(1, $a->getDependencies());
         $a->addDependency($b);
-        $this->assertCount(1, $a->getDependencies());
+        self::assertCount(1, $a->getDependencies());
     }
 
-    public function testHasDependency()
+    public function testHasDependency(): void
     {
         $a = new Job('a');
         $b = new Job('b');
 
-        $this->assertFalse($a->hasDependency($b));
+        self::assertFalse($a->hasDependency($b));
         $a->addDependency($b);
-        $this->assertTrue($a->hasDependency($b));
+        self::assertTrue($a->hasDependency($b));
     }
 
-    public function testIsRetryAllowed()
+    public function testIsRetryAllowed(): void
     {
         $job = new Job('a');
-        $this->assertFalse($job->isRetryAllowed());
+        self::assertFalse($job->isRetryAllowed());
 
         $job->setMaxRetries(1);
-        $this->assertTrue($job->isRetryAllowed());
+        self::assertTrue($job->isRetryAllowed());
 
         $job->setState('running');
         $retry = new Job('a');
         $job->addRetryJob($retry);
-        $this->assertFalse($job->isRetryAllowed());
+        self::assertFalse($job->isRetryAllowed());
     }
 
-    public function testCloneDoesNotChangeQueue()
+    public function testCloneDoesNotChangeQueue(): void
     {
         $job = new Job('a', array(), true, 'foo');
         $clonedJob = clone $job;
 
-        $this->assertEquals('foo', $clonedJob->getQueue());
+        self::assertEquals('foo', $clonedJob->getQueue());
     }
 
-    private function setField($obj, $field, $value)
+    private function setField($obj, $field, $value): void
     {
         $ref = new \ReflectionProperty($obj, $field);
         $ref->setAccessible(true);
