@@ -14,7 +14,7 @@ class ConcurrencyTest extends BaseTestCase
     private $configFile;
     private $databaseFile;
 
-    public function testHighConcurrency()
+    public function testHighConcurrency(): void
     {
         $this->startWorker('one');
         $this->startWorker('two');
@@ -53,7 +53,7 @@ class ConcurrencyTest extends BaseTestCase
         $this->assertEquals(array('one', 'two'), $workers);
     }
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $this->databaseFile = tempnam(sys_get_temp_dir(), 'db');
         $this->configFile = tempnam(sys_get_temp_dir(), 'di-cfg');
@@ -77,7 +77,7 @@ CONFIG
         $this->importDatabaseSchema();
     }
 
-    protected function tearDown()
+    protected function tearDown(): void
     {
         @unlink($this->databaseFile);
         @unlink($this->configFile);
@@ -91,7 +91,7 @@ CONFIG
         }
     }
 
-    private function waitUntilJobsProcessed($maxRuntime)
+    private function waitUntilJobsProcessed($maxRuntime): void
     {
         $start = time();
         do {
@@ -114,7 +114,7 @@ CONFIG
         }
     }
 
-    private function startWorker($name)
+    private function startWorker($name): void
     {
         $proc = new Process('exec '.PHP_BINARY.' '.escapeshellarg(__DIR__.'/console').' jms-job-queue:run --worker-name='.$name, null, array(
             'SYMFONY_CONFIG' => $this->configFile,
